refactor: separate admin actions on product from customer actions on product, apply authentication on product endpoints, add id to product dto, change rating to double instead of int
