+++ src/main/java/com/academy/e_commerce/controller/ProductPortalController.java
@@ -1,0 +1,34 @@
+package com.academy.e_commerce.controller;
+
+import com.academy.e_commerce.dto.ProductDTO;
+import com.academy.e_commerce.model.Product;
+import com.academy.e_commerce.service.ProductService;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+@RequestMapping("${api.endpoint.base-url}/portal/products")
+public class ProductPortalController {
+
+    private final ProductService productService;
+
+    public ProductPortalController(ProductService productService) {
+        this.productService = productService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Product> createProduct(@RequestBody ProductDTO productDTO) {
+        return ResponseEntity.ok(productService.createProduct(productDTO));
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Product> updateProduct(@PathVariable Long id, @RequestBody ProductDTO productDTO) {
+        return ResponseEntity.ok(productService.updateProduct(id, productDTO));
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteProduct(@PathVariable Long id) {
+        productService.deleteProduct(id);
+        return ResponseEntity.noContent().build();
+    }
+}
