+++ src/main/java/com/academy/e_commerce/dto/ProductDTO.java
@@ -2,11 +2,12 @@ package com.academy.e_commerce.dto;
 
 
 public record ProductDTO(
+        Long id,
         String name,
         String description,
         Integer stock,
         Double price,
         String category,
         String image,
-        int rating
+        Double rating
 ) {}
